requirements.txt
transformers
flask
gunicorn
pillow
requests